+++ tests/functional/test_custom_outputs.py
@@ -84,4 +84,4 @@ def test_list_output():
     mean = np.mean(sample.samples['p'], axis=0)
 
     # Crude test
-    assert mean[1] - mean[0] > .2
+    assert mean[1] > mean[0]