+++ CHANGELOG.rst
@@ -5,7 +5,7 @@ Changelog
 dev
 ---
 
-- Reusing a pool with a new inference
+- Fix a bug preventing the reuse of ArrayPool data with a new inference
 
 
 0.6 (2017-07-03)